% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{AIC}
\alias{AIC}
\title{Calculate AIC based on MSE.}
\usage{
AIC(MSE = NULL, logLik = NULL, k, N)
}
\arguments{
\item{MSE}{A vector of Mean Squared Errors, one for each model.}

\item{logLik}{A vector of log-likelihoods, one for each model.}

\item{k}{A vector of the number of free parameters for each model.}

\item{N}{The number of observations in the data set.}
}
\value{
A vector of AIC values for each model.
}
\description{
This function calculates a vector of AIC values for a set of models, 
based on either their mean squared errors (MSE) or their log-likelihoods. 
The AIC is a measure of the relative quality of statistical models for a given set
of data, with lower values indicating better model fit.
}
\details{
Either a vector of MSE values or a vector of log-likelihoods must be provided,
along with the number of free parameters (k) for each model and the number of observations 
(N) in the data set.

It is customary (e.g. in other functions here) to minimize the _negative_ log-likelihood, 
so that lower values indicate better fits. Make sure to multiply it any _negative_ 
log-likelihoods by -1 for use in this function.
}
\examples{
#
}
