% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scrubbing.R
\name{circleCorrect}
\alias{circleCorrect}
\title{Shift localization responses so they are centred on the origin}
\usage{
circleCorrect(df)
}
\arguments{
\item{df}{Data frame of localization data}
}
\value{
The data frame with corrected \code{tapx_cm} and \code{tapy_cm} columns. The
corrected localization responses fall closest to a circle with radius 12 cm and
origin (0,0). Only response with \code{df$selected == 1} are used for this correction.
}
\description{
Shift localization responses so they are centred on the origin
}
