% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scrubbing.R
\name{circleErrors}
\alias{circleErrors}
\title{Get mean squared error between coordinates and a circle}
\usage{
circleErrors(par, X, Y, r)
}
\arguments{
\item{par}{Vector with xc and yc parameters: x and y of the circle's middle}

\item{X}{Vector of X coordinates}

\item{Y}{Vector of Y coordinates}

\item{r}{The radius of the circle}
}
\value{
The mean squared error between the distances of \code{X} and 
\code{Y} from the position in par and the radius \code{r}.
}
\description{
Get mean squared error between coordinates and a circle
}
\examples{

}
