% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{colorAlpha}
\alias{colorAlpha}
\title{Set or change the alpha (transparency) of a color}
\usage{
colorAlpha(col, alpha = 34)
}
\arguments{
\item{col}{Color}

\item{alpha}{Alpha value to apply}
}
\value{
An RGBA color
}
\description{
Sets the alpha of an RGB(A) color
}
