% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{colorMix}
\alias{colorMix}
\title{Mix two colors}
\usage{
colorMix(a = "#41ffc9", b = "#1d7791", balance = c(1, 1))
}
\arguments{
\item{a}{First color}

\item{b}{Second color}

\item{balance}{Numeric vector of length 2, Weights of the two colors}
}
\value{
A weighted average of the two colors in RGB space
}
\description{
Mixes two colors in RGB space using weights
}
