% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{colorSaturation}
\alias{colorSaturation}
\title{Change the saturation of a color}
\usage{
colorSaturation(col = "#41ffc9", sat.mult = 1.25)
}
\arguments{
\item{col}{Color}

\item{sat.mult}{Multiply saturation with this number}
}
\value{
The input color with the saturation multiplied.
}
\description{
Converts the input color to HSV space, then multiplies the saturation
by the saturation multiplier.
}
