% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert2RData.R
\name{convert2RData}
\alias{convert2RData}
\title{Converts an eligible Matlab .mat file to an .RData file}
\usage{
convert2RData(matfile)
}
\arguments{
\item{matfile}{A string denoting the matfile, e.g. "mymatlabdata" or "mymatlabdata.mat"}
}
\description{
Converts an eligible Matlab .mat file to an .RData file. Keeps the .mat file unchanged.
The Matlab file must have been saved with the -v7 option flag.
}
\examples{
v <- sample(1:10,4)
m <- matrix(runif(9),3,3)
print(v)
print(m)
R.matlab::writeMat("test_convert2RData.mat", v=v, m=m)
rm(m,v)
convert2RData("test_convert2RData.mat")
load("test_convert2RData.RData")
print(v)
print(m)
}
\author{
Christoph Schmidt <christoph.schmidt@med.uni-jena.de>
}
\seealso{
\code{\link{runMatlabScript}}, \code{\link{runMatlabCommand}}
}

