% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSFdata.R, R/data_syncing.R
\name{dataFilecheck}
\alias{dataFilecheck}
\title{Check if csv files from requested groups and sections exist}
\usage{
dataFilecheck(groups = "all", sections = "sections", verbose = FALSE)

dataFilecheck(groups = "all", sections = "sections", verbose = FALSE)
}
\arguments{
\item{groups}{vector of group names for which to download data (default: 'all')}

\item{sections}{vector of task sections for which to download data (default: 'all')}

\item{verbose}{boolean, if \code{TRUE}: print names of all missing files}

\item{groups}{vector of group names for which to download data (default: 'all')}

\item{sections}{vector of task sections for which to download data (default: 'all')}

\item{verbose}{boolean, if \code{TRUE}: print names of all missing files}
}
\value{
empty

empty
}
\description{
Check if csv files from requested groups and sections exist

Check if csv files from requested groups and sections exist
}
\examples{
#dataFilecheck(groups='all', sections=c('aligned'), verbose=TRUE)
#dataFilecheck(groups='all', sections=c('aligned'), verbose=TRUE)
}
