% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df.col.grid}
\alias{df.col.grid}
\title{Expand grid with values from data frame columns}
\usage{
df.col.grid(df, columns)
}
\arguments{
\item{df}{A data frame.}

\item{columns}{A list of column names in the data frame.}
}
\value{
A data frame with all combinations of values existing in the data frame
on the specified columns.
}
\description{
This is similar to expand.grid but uses the existing values in a
data frame, on the variables that you name. This would be most useful as a step
in preparing data for some other analysis.
}
\details{
Not yet.
}
\examples{
#
}
