% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_syncing.R
\name{downloadData}
\alias{downloadData}
\title{Download data from OSF and unzip into folder structure in \code{data/} folder.}
\usage{
downloadData(groups = "all", sections = "all", overwrite = FALSE)
}
\arguments{
\item{groups}{vector of group names for which to download data (default: 'all')}

\item{sections}{vector of task sections for which to download data (default: 'all')}

\item{repository}{OSF repository to download from}
}
\value{
empty
}
\description{
Download data from OSF and unzip into folder structure in \code{data/} folder.
}
\examples{
#dataDownload(groups=c('control', 'instructed'), sections=c('aligned'))
}
