% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSFdata.R
\name{downloadOSFdata}
\alias{downloadOSFdata}
\title{Download data from OSF and unzip into folder structure in \code{data/} folder.}
\usage{
downloadOSFdata(repository, filelist, folder, overwrite = TRUE)
}
\arguments{
\item{repository}{OSF repository to download from (string of 5 characters)}

\item{filelist}{named list where names are folders, and entries are vectors
of filenames to download from those folders on the OSF repository}

\item{folder}{folder in the current working directory to store files in}
}
\value{
empty
}
\description{
Download data from OSF and unzip into folder structure in \code{data/} folder.
}
