% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{dprime}
\alias{dprime}
\title{Calculate d-prime and related measures from signal detection data.}
\usage{
dprime(X, hits = NULL, misses = NULL, fas = NULL, crs = NULL, hautus = FALSE)
}
\arguments{
\item{X}{A data frame with two columns: first column with signal present (1)
or absent (0), second column with response yes (1) or no (0).}

\item{hits}{Number of hits.}

\item{misses}{Number of misses.}

\item{fas}{Number of false alarms.}

\item{crs}{Number of correct rejections.}

\item{hautus}{Boolean: use Hautus correction for extreme values (default: FALSE).}
}
\value{
A list with d-prime, beta, c, sensitivity and specificity.
}
\description{
Calculate d-prime and related measures from signal detection data.
}
\details{
This function calculates d-prime, beta, c, sensitivity and specificity
from counts of hits, misses, false alarms and correct rejections.
It can also apply the Hautus correction for extreme values (0 or 1 rates).
If X is provided as a data frame, the counts are calculated from that, and the 
other four parameters are ignored.
}
\examples{
#' results <- dprime(hits=45, misses=5, fas=15, crs=35, hautus=TRUE)
#' print(results)
}
