% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{exponentialModel}
\alias{exponentialModel}
\title{Run an exponential function given parameters and a reach
deviation schedule. Errors decay exponentially.}
\usage{
exponentialModel(par, timepoints, mode = "learning")
}
\arguments{
\item{par}{A named vector with the model parameter (see details).}

\item{timepoints}{An integer indicating the number of trials (N), or a vector
with N trial numbers (these can have missing values or be fractions). If an
integer, the timepoints at which the exponential will be evaluated is:
0, 1 ... N-2, N-1}

\item{mode}{String: "learning" or "washout", sets the function's direction.}
}
\value{
A data frame with two columns: `timepoint` and `output`, and N rows,
so that each row has the output of the modeled process on each trial.
}
\description{
This function is part of a set of functions to fit and
evaluate an exponential decay model with asymptote.
}
\details{
The `par` argument is a named numeric vector that should have the
following element:
- lambda: learning rate
- N0: asymptote
}
\examples{
# write example!
}
