% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitProbDens}
\alias{fitProbDens}
\title{Fit a probability density function to a vector of data.}
\usage{
fitProbDens(x, densfun = "normal")
}
\arguments{
\item{x}{Vector of data points}

\item{densfun}{String indicating the type of density function to fit. Options are:
- 'normal': Normal distribution
- 'lognormal': Log-normal distribution
- 'gamma': Gamma distribution
- 'beta': Beta distribution}
}
\description{
This function fits a specified probability density function to a vector
of data points. It uses maximum likelihood estimation to find the best-fitting 
parameters for the chosen distribution.
}
\details{
The function supports fitting the following distributions:
- Normal distribution: Parameters are mean and standard deviation.
- Log-normal distribution: Parameters are meanlog and sdlog.
- Gamma distribution: Parameters are shape and scale.
- Beta distribution: Parameters are shape1 and shape2.
}
\examples{
# 
}
