% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get95CIellipse}
\alias{get95CIellipse}
\title{Calculate surface area of 95% confidence ellipse on 2D data.}
\usage{
get95CIellipse(df, vars = NULL)
}
\arguments{
\item{df}{A data frame, with (at least) 2 variables with the same unit.}

\item{vars}{A list of column to use in the data frame, if `df` has more than 
2 columns.}
}
\value{
The surface area of an ellipse that contains the mean of the data in 95%
of replications of the same experiment.
}
\description{
This function uses PCA to calculate the standard deviation across
the two main axes of the spread of 2D data. Multiplying that length-2 vector 
by `pi` and `qnorm(0.975)` gives the surface area of an ellipse that will 
contain the mean in 95% of cases.
This relies on the data being in the same unit, and makes most sense if it is real 
data (such as locations in a cartesian coordinate system).
}
\details{
Not yet.
}
