% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{getColors}
\alias{getColors}
\title{Get colors that are commonly used in the lab}
\usage{
getColors()
}
\value{
A list of two vectors of colors. The list has two elements:
`op` and `tr`. `op` contains opaque colors, `tr` contains the same colors but
with transparency (alpha = 32).
}
\description{
This function returns a list of two vectors of colors.
The list has two elements:
`op` and `tr`. `op` contains opaque colors, `tr` contains the same colors but
with transparency (alpha = 32). These colors are commonly used in the lab for
figures on posters and in papers. The second color is a red, which is taken 
from the York logo. The colors are sorted from most to least bright.
}
