% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{getLocalizationSpline}
\alias{getLocalizationSpline}
\title{Get a smoothed spline object fit on localization errors over reach angles.}
\usage{
getLocalizationSpline(df, spar = 0.5)
}
\arguments{
\item{df}{Data frame with localization data, this contains at least the
variables: \code{reachangle_deg} and \code{localizationerror_deg}.}

\item{spar}{Smoothing parameter for the spline (0,1) (default: \code{0.50}).}
}
\value{
A spline object that predicts localization errors over reach angle.
}
\description{
Get a smoothed spline object fit on localization errors over reach angles.
}
