% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatlabCall.R
\name{getMatlabCall}
\alias{getMatlabCall}
\title{Get OS-specific Matlab program call string}
\usage{
getMatlabCall()
}
\description{
Returns the matlab programm name for OS X, Linux and Windows.
}
\author{
Christoph Schmidt <christoph.schmidt@med.uni-jena.de>
}

