% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWin.R
\name{isWin}
\alias{isWin}
\title{Is the underlying OS Windows?}
\usage{
isWin()
}
\description{
Returns a logical that indicates whether \pkg{reach} is used on a Windows
machine (which is useful information for adjusting path specifications).
}
\author{
Christoph Schmidt <christoph.schmidt@med.uni-jena.de>
}

