% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{latinSquare}
\alias{latinSquare}
\title{Latin Square}
\usage{
latinSquare(order, seed = NULL)
}
\arguments{
\item{order}{An integer value of 2 or higher, setting the size of the latin square.}

\item{seed}{Optional seed for the random number generator}
}
\value{
A matrix of size order x order
}
\description{
Returns a latin square where each index 1:order occurs in each row
as well as in each column.
}
\details{
If you want the same latin square repeatedly, set the seed argument.
}
\examples{
latinSquare(5)
}
