% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{localizationSD}
\alias{localizationSD}
\title{Get the standard deviation of localization using a spline as 'mean'.}
\usage{
localizationSD(df, unit = "cm", locvar = "tap", handvar = "hand",
  r = 1, CC = TRUE, spar = 0.99, rm.Extr = FALSE)
}
\arguments{
\item{df}{Data frame with localization data.}

\item{unit}{The unit of position data (default: \code{cm}).}

\item{locvar}{The variable name of localization responses (default: \code{tap}).}

\item{handvar}{The variable name of hand position (default: \code{hand}).}

\item{r}{Radius of localization arc (default: 12).}

\item{CC}{(boolean) Perform circle correction (default: \code{TRUE}).}

\item{spar}{Smoothing parameter for the spline (0,1) (default: \code{0.99}).}

\item{rm.Extr}{(boolean) Remove the highest and lowest reach angles before 
spline interpolation (after fitting the spline).}
}
\value{
A spline object that predicts localization errors over reach angle.
}
\description{
Get the standard deviation of localization using a spline as 'mean'.
}
