% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{multiModalModel}
\alias{multiModalModel}
\title{Probability of data points given a multi-modal distribution.}
\usage{
multiModalModel(par, x)
}
\arguments{
\item{par}{A named vector with the model parameter (see details).}

\item{x}{A sequence of numbers to evaluate the probability of.}
}
\value{
A vector of probabilities according to the multi modal normal distribution.
}
\description{
This function is part of a set of functions to fit and
evaluate multi modal (normal) distribution of data points.
}
\details{
The `par` argument is a data frame (or named list) with columns:
- m: mean
- s: standard deviation
- w: weight
for every normal distribution of the model
}
