% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{multiModalModelLikelihood}
\alias{multiModalModelLikelihood}
\title{Likelihood of data given a multi-modal probability distribution.}
\usage{
multiModalModelLikelihood(par, x)
}
\arguments{
\item{par}{A named vector with the model parameter (see details).}

\item{x}{A sequence of numbers to evaluate the probability of.}
}
\value{
The sum of the log of the probabilities returned by `multiModalModel()`.
}
\description{
This function is part of a set of functions to fit and
evaluate multi modal (normal) distribution of data points.
}
\details{
The `par` argument is a data frame (or named list) with columns:
- m: mean
- s: standard deviation
- w: weight
for every normal distribution of the model
}
