% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nll}
\alias{nll}
\title{negative log-likelihood.}
\usage{
nll(d)
}
\arguments{
\item{d}{Vector of density values for observations given a model.}
}
\description{
This function calculates the negative log-likelihood of a set of density
values. It is used to evaluate how well a model fits the data, with lower values 
indicating a better fit.
}
\details{
The input vector `d` should contain the density values for each observation 
given a model. The function will replace any density values that are exactly 1 with 
a very small value to avoid taking the log of 0, which would result in an infinite 
negative log-likelihood. If the resulting negative log-likelihood is not finite, it 
will be set to the minimum possible value based on the number of observations.
}
\examples{
# 
}
