% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ODR}
\alias{ODR}
\title{Orthogonal Distance Regression.}
\usage{
ODR(X, y)
}
\arguments{
\item{X}{A predictor matrix (variables in columns, N observations in rows).}

\item{y}{Column vector (N rows) with dependent variable.}
}
\value{
A list with: `beta` (row vector of weights) and `intercept`.
}
\description{
Perform orthogonal distance regression with multiple predictors.
}
\details{
This function uses PCA to perform orthogonal distance regression
or "total least squares regression") on y based on one or more predictors
in X.
Code taken from https://stats.stackexchange.com/questions/13152/how-to-perform-orthogonal-regression-total-least-squares-via-pca
}
\examples{
# we'll predict petal width from other iris virginica properties:
y <- iris$Petal.Width[101:150]
X <- as.matrix(iris[101:150,1:3])
pw.odr <- ODR(X,y)

# let's have a look at how well that predicts petal width:
y_hat <- (X \%*\% pw.odr$beta) + pw.odr$intercept
plot(y,y_hat,asp=1,xlim=c(0,3),ylim=c(0,3))
}
