% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{offsetErrorDecayModel}
\alias{offsetErrorDecayModel}
\title{Values of an offset exponential decay given parameters and time points}
\usage{
offsetErrorDecayModel(par, timepoints)
}
\arguments{
\item{par}{A named vector with the function parameter (see details).}

\item{timepoints}{An integer indicating the number of trials (N), or a vector
with N trial numbers (these can have missing values or be fractions) with the
first timepoint having index 0.
If this is an integer, the timepoints at which the exponential will be 
evaluated is: 0, 1 ... N-2, N-1}
}
\value{
A data frame with two columns: `trial` and `value`, and N rows,
so that each row has the output of the modeled process on each trial.
}
\description{
This function is part of a set of functions to fit and
evaluate an exponential decay model with asymptote and offset.
}
\details{
The `par` argument is a named numeric vector that should have the
following elements:
- r: rate (of decay/learning)
- s: span (difference between starting point and asymptote)
- o: offset (added to the whole function)
}
