% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{offsetErrorDecayMSE}
\alias{offsetErrorDecayMSE}
\title{Get the MSE between offset exponential decay and a time series}
\usage{
offsetErrorDecayMSE(par, signal, timepoints = c(0:(length(signal) - 1)))
}
\arguments{
\item{par}{A named numeric vector with the model parameters (see details).}

\item{signal}{A numeric vector of length N with reach deviations matching
the perturbation schedule.}

\item{timepoints}{Either an integer with the number of trials (N) or a vector
with N trial numbers (this can have missing values or fractions). The 
exponential will be evaluated at those timepoints.}
}
\value{
A float: the mean squared error between the total model output and
the time series.
}
\description{
This function is part of a set of functions to fit and
evaluate an offset exponential decay function to a time series.
}
\details{
The `par` argument is a named numeric vector that should have the
following elements:
- r: rate (of decay/learning)
- s: span (difference between starting point and asymptote)
- o: offset (added to the whole function)
}
