% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{posterGuides}
\alias{posterGuides}
\title{Calculate coordinates for guides on a poster}
\usage{
posterGuides(size = 215.9, outermargins = 10, innermargins = 5, ncolumns = 3)
}
\arguments{
\item{size}{The size (usually width) of the poster}

\item{outermargins}{The width of the outer margins}

\item{innermargins}{The width of the margins in between columns}

\item{ncolumns}{The number of columns that should appear on the poster}
}
\value{
a vector of coordinates for guides
}
\description{
Given the size (width) of a poster, the size of the inner and
outer margins and the number of columns, this function returns the location
of guides. 

All arguments except `ncolumns` should be in the same unit.
}
\examples{
# the default values use the width of a landscape 'letter' in mm
# with outer margins of 10 mm, and inner margins of 5 mm
# and 3 columns
posterGuides()
}
