% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{prepareSplineVariables}
\alias{prepareSplineVariables}
\title{Prepare data for smoothed spline fitting.}
\usage{
prepareSplineVariables(
  df,
  CC = TRUE,
  handvar = "hand",
  locvar = "tap",
  unit = "cm",
  r = 1
)
}
\arguments{
\item{df}{Data frame with localization data, this contains at least the
variables: \code{reachangle_deg} and \code{localizationerror_deg}.}

\item{CC}{(boolean) Apply circleCorrect or not (default: \code{TRUE}).}

\item{handvar}{Part of column names specifying reach endpoints (default: \code{hand}).}

\item{locvar}{Part of column names specifying hand location (default: \code{tap}).}

\item{unit}{Unit of the four location variables (default: \code{cm}).}

\item{r}{Radius for the circleCorrect procedure (default: \code{1})}
}
\value{
Hand lcalization data frame extended with columns to be used for
smooth spline fitting.
}
\description{
Prepare data for smoothed spline fitting.
}
