% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.R
\name{reach}
\alias{reach}
\title{Regularized rank-constrained heterogeneity pursuit (REACH)}
\usage{
reach(
  Y,
  X,
  Z_,
  U0 = NULL,
  V0 = NULL,
  W0 = NULL,
  penalty = c("MCP", "SCAD", "L1"),
  lambda.B = NULL,
  lambda.W = NULL,
  r = 3,
  rho = 1,
  a = 3,
  epsilon = 0.001
)
}
\arguments{
\item{Y}{response matrix}

\item{X}{covariate matrix}

\item{Z_}{control variable matrix}

\item{U0}{initial value of U; if NULL, generated by reduced-rank regression}

\item{V0}{initial value of V; if NULL, generated by reduced-rank regression}

\item{W0}{initial value of W; if NULL, generated by reduced-rank regression}

\item{penalty}{select penalty function for heterogeneity exploration; the default is MCP}

\item{lambda.B}{sequence of tuning parameters for B; if NULL, generated by a data driven procedure}

\item{lambda.W}{sequence of tuning parameters for W; if NULL, generated by a data driven procedure}

\item{r}{an integer specifying the desired rank; the default is 3}

\item{rho}{penalty parameter in augmented Lagrangian; the default is 1}

\item{a}{parameter that controls the concavity of penalty function; only useful for concave penalties; the default is 3}

\item{epsilon}{tolerance in algorithm; the default is 1e-3}
}
\value{
A \code{REACH} object containing
\item{call}{original function call}
\item{Y}{input response matrix}
\item{X}{input covariate matrix}
\item{Z_}{input control variable matrix}
\item{B}{estimated coefficient matrix for X}
\item{W}{estimated coefficient matrix for Z}
\item{Theta}{a list of estimated common value matrices}
\item{C}{estimated pairwise difference matrix}
\item{s}{estimated sparsity level}
\item{p.index}{indices of covariates which passed screening}
\item{r}{estimated rank}
\item{group}{a vector of estimated group labels}
\item{K}{estimated group number}
\item{lambda.B}{selected tuning parameter for B}
\item{lambda.W}{selected tuning parameter for W}
\item{GIC}{GIC value for the fitted model}
\item{iter}{iteration number}
\item{res}{final residual in algorithm}
}
\description{
Compute solutions of regularized rank-constrained heterogeneity pursuit
}
\examples{
library(reach)
 ## Simulate data from a heterogeneous sparse reduced-rank regression model
 p <- 100; m <- 1; q <- 50; n <- 100; r <- 3; s = 5
 mydata <- reach.sim(n, p, q, m, r, s, rho_X = .5, sigma_X = 1, s2n = 1)

 Y <- mydata$Y
 X <- mydata$X
 Z_ <- mydata$Z_

 fit <- reach(Y, X, Z_, penalty = "MCP")

 fit$s
 fit$K
 fit$B
 fit$Theta

}
