% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.sim.R
\name{reach.sim}
\alias{reach.sim}
\title{Simulate heterogeneous sparse reduced-rank regressions}
\usage{
reach.sim(n, p, q, m, r, s, rho_X = 0.5, sigma_X = 1, s2n = 1)
}
\arguments{
\item{n, p, m, q}{model dimensions}

\item{r}{model rank}

\item{s}{model sparsity level}

\item{rho_X}{correlation parameter in the generation of covariates}

\item{sigma_X}{variance parameter in the generation of covariates}

\item{s2n}{signal to noise ratio}
}
\value{
similated model and data
}
\description{
Similar to the models in simulation examples 1-2 in Li et al. (2025)
}
