% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{relativeLikelihood}
\alias{relativeLikelihood}
\title{Calculate relative likelihood.}
\usage{
relativeLikelihood(crit)
}
\arguments{
\item{crit}{Vector of criterion values for models.}
}
\value{
Vector of likelihoods. The best model will have a likelihood of 1, and
models with likelihoods lower than 0.05 could be considered "significantly" worse
than the best model.
}
\description{
This function calculate a relative log likelihood using criterion values
for a set of models. The criterion should indicate a better model for a lower value.
}
\details{
If the input vector is named, the output vector will also be named.
}
\examples{
#
}
