% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runMatlabCommand.R
\name{runMatlabCommand}
\alias{runMatlabCommand}
\title{Starts Matlab on the R console and executes the input command}
\usage{
runMatlabCommand(commandName)
}
\arguments{
\item{commandName}{String denoting the matlab command}
}
\description{
Starts Matlab on the R console and executes the input command and quits Matlab.
Discerns the Mac and Linux Matlab command.
}
\details{
As R and Matlab cannot directly exchange data natively, no value can be returned.
Instead, let Matlab save the results of its computations and load these into R for further processing.
An error in the Matlab command prevents Matlab from quitting in the R console and might
require an re-start of the R session. So check the command in Matlab before executing it within R.
The commandName could look something like this:
"'load someData.mat; [ca,Q]=modularity_dir(A); save someData2.mat ca Q; quit'"
}
\examples{
\dontrun{

commandName <- "'x=1:2:7; y=3; disp(x.^y); quit'"
runMatlabCommand(commandName)
}
}
\author{
Christoph Schmidt <christoph.schmidt@med.uni-jena.de>
}
\seealso{
\code{\link{convert2RData}}, \code{\link{runMatlabScript}}
}

