% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{setupFigureFile}
\alias{setupFigureFile}
\title{Set up a graphics device for saving figures}
\usage{
setupFigureFile(target = "inline", width = 8, height = 6, dpi = 300, filename)
}
\arguments{
\item{target}{The type of file to create. One of 'pdf', 'svg', 'png', 'tiff'
or 'inline'. Default is 'inline', which does not open a file, but leaves use
to the default device (e.g. RStudio plot pane).}

\item{width}{Width of the figure in inches. Default is 8.}

\item{height}{Height of the figure in inches. Default is 6.}

\item{dpi}{Dots per inch (resolution) for raster formats (png, tiff). Default is 300.}

\item{filename}{The name of the file to create, including the path. Required if 
target is not 'inline'.}
}
\value{
Does not return a value.
}
\description{
Boiler plate code to set up a graphics device for saving figures.
The function does nothing is the target is not one of the 4 file formats allowed,
such that the it can be both used for creating publication quality files and to
plot to interactive graphics devices (i.e. on the screen).
}
