% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_extractCommaSepArgs.R
\name{str_extractCommaSepArgs}
\alias{str_extractCommaSepArgs}
\title{Extract arguments given in a string, separated by commas}
\usage{
str_extractCommaSepArgs(args_str)
}
\arguments{
\item{args_str}{character vector containing (variable) names separated by
  commas}
}
\value{
Returns a list containing the (variable) names (characters) stored in
  \code{args_str}.
}
\description{
Parses an input string and returns the comma separated arguments.
}
\examples{
vars    <- "A, B, myChar, level, k,  test, te_st  , newval"
reach:::str_extractCommaSepArgs(vars)
}
\author{
Christoph Schmidt <christoph.schmidt@med.uni-jena.de>
}

