% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSFdata.R
\name{unzipZips}
\alias{unzipZips}
\title{Download data from OSF and unzip into folder structure in \code{data/} folder.}
\usage{
unzipZips(filelist, folder, removezips = FALSE)
}
\arguments{
\item{filelist}{named list where names are folders, and entries are vectors
of filenames to download from those folders on the OSF repository}

\item{folder}{folder in the current working directory to store files in}

\item{removezips}{(Boolean) remove zip files after unzipping?}
}
\value{
empty
}
\description{
Download data from OSF and unzip into folder structure in \code{data/} folder.
}
