% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urls}
\alias{urls}
\title{URLs to download the data from OSF}
\format{A data frame with 38 rows and 3 variables:
\describe{
  \item{group}{one of 14 experimental groups}
  \item{section}{one of 3 sections of data}
  \item{URL}{the urls of the corresponding zip file}
}}
\usage{
urls
}
\description{
Description
}
\keyword{datasets}
