% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{AIC}
\alias{AIC}
\title{Calculate AIC based on MSE.}
\usage{
AIC(MSE, k, N)
}
\arguments{
\item{MSE}{A vector of Mean Squared Errors, one for each model.}

\item{k}{A vector of the number of free parameters for each model.}

\item{N}{The number of observations in the data set.}
}
\value{
A vector of AIC values for each model.
}
\description{
This function
}
\details{
#
}
\examples{
#
}
