% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{circleCorrect}
\alias{circleCorrect}
\title{Shift localization responses so they are centred on the origin}
\usage{
circleCorrect(df, unit = "cm", vrbl = "tap", r = 1, fitr = FALSE)
}
\arguments{
\item{df}{Data frame with localization coordinates (X,Y).}

\item{unit}{Unit of the coordinates (default: 'cm')}

\item{r}{Radius of the circle the coordinates should be on (default: 1).}

\item{fitr}{(boolean) Should radius be fit? (default: FALSE)}

\item{var}{Variable of coordinates (default: 'tap')}
}
\value{
The data frame with corrected \code{tapx_cm} and \code{tapy_cm} 
columns. The corrected localization responses fall closest to a circle with
radius \code{r} (in \code{unit}) and origin (0,0). Only response with
\code{df$selected == 1} are used for this correction.
}
\description{
Shift localization responses so they are centred on the origin
}
\details{
The parameters \code{var} and \code{unit} are combined with a lower
case \code{x} and \code{y}: \code{tapx_cm} and \code{tapy_cm} with default
settings. These should be columns in the data frame (\code{df}).
}
