% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert2RData.R
\name{convert2RData}
\alias{convert2RData}
\title{Converts an eligible Matlab .mat file to an .RData file}
\usage{
convert2RData(matfile = NULL, dir = NULL)
}
\arguments{
\item{matfile}{string or string vector denoting one or several matfiles, e.g.
  "mymatlabdata", "mymatlabdata.mat", "path/to/mymatlabfile.mat" or
  c("one.mat", "two.mat")}

\item{dir}{path to a directory which contains one or several .mat files that
  should be converted to .RData files}
}
\description{
Converts an eligible Matlab .mat file in the current directory to an .RData
file. Keeps the .mat file unchanged. The Matlab file must have been saved
with the -v7 option flag. Can also convert either all .mat files contained in
a specified directory or a single specified .mat file in a specified
directory.
}
\details{
If single .mat files in the current directory should be converted to
  .RData, then the input argument dir has to be set to NULL. If the input
  argument matfile is NULL and the input argument dir is specified, then all
  .mat files in the given directory will be converted to .RData. If on the
  other hand both input arguments, dir and matfile, are specified, then the
  given .mat file(s) in the given directory will be converted to .RData.
  Providing the file type specifier ".mat" is optional.
}
\examples{
\dontrun{

##### conversion of a single .mat file in the current working directory ####
v <- sample(1:10,4)
m <- matrix(runif(9),3,3)
print(v)
print(m)
R.matlab::writeMat("file_convert2RData.mat", v=v, m=m)
rm(v,m)
print(ls())

convert2RData("./file_convert2RData.mat")

load("file_convert2RData.RData")
print(v)
print(m)
file.remove(c("file_convert2RData.RData", "file_convert2RData.mat"))



#### conversion of all .mat files in a specified directory ####
this_dir <- getwd()
m   <- matrix(runif(9),3,3)
v   <- seq(1,100)
print(v)
print(m)
R.matlab::writeMat("dir_convert2RData_1.mat", m=m)
R.matlab::writeMat("dir_convert2RData_2.mat", v=v)
rm(v,m)

convert2RData(dir=this_dir)

load("dir_convert2RData_1.Rdata")
print(m)
load("dir_convert2RData_2.Rdata")
print(v)
file.remove(c("dir_convert2RData_1.mat", "dir_convert2RData_2.mat",
   "dir_convert2RData_1.Rdata", "dir_convert2RData_2.Rdata"))



#### conversion of a single specified .mat file in a specified directory ####
this_dir <- getwd()
v   <- seq(1,10)
print(v)
R.matlab::writeMat("file_dir_convert2RData.mat", v=v)
rm(v)

convert2RData("file_dir_convert2RData.mat", this_dir)

load("file_dir_convert2RData.RData")
print(v)
file.remove(c("file_dir_convert2RData.mat", "file_dir_convert2RData.RData"))



#### conversion of several specified .mat files in the current working directory ####
v <- sample(1:10,4)
m <- matrix(runif(9),3,3)
print(v)
print(m)
R.matlab::writeMat("twofiles_convert2RData_1.mat", v=v)
R.matlab::writeMat("twofiles_convert2RData_2.mat", m=m)
rm(v,m)

convert2RData(c("twofiles_convert2RData_1.mat", "twofiles_convert2RData_2.mat"))

load("twofiles_convert2RData_1.RData")
print(v)
load("twofiles_convert2RData_2.RData")
print(m)
file.remove(c("twofiles_convert2RData_1.mat", "twofiles_convert2RData_2.mat",
   "twofiles_convert2RData_1.RData", "twofiles_convert2RData_2.RData"))
}
}
\author{
Christoph Schmidt <christoph.schmidt@med.uni-jena.de>
}
\seealso{
\code{\link{runMatlabScript}}, \code{\link{runMatlabCommand}}
}

