% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{exponentialMSE}
\alias{exponentialMSE}
\title{Get the MSE between an exponential and a series of reach deviations.}
\usage{
exponentialMSE(par, signal, timepoints = NULL)
}
\arguments{
\item{par}{A named numeric vector with the model parameters (see
exponentialModel).}

\item{signal}{A numeric vector of length N with reach deviations matching
the perturbation schedule.}

\item{timepoints}{Either an integer with the number of trials (N) or a vector
with N trial numbers (this can have missing values or fractions). The 
exponential will be evaluated at those timepoints.}
}
\value{
A float: the mean squared error between the total model output and
the reach deviations.
}
\description{
This function is part of a set of functions to fit and
evaluate an exponential function to describe a series of reach deviations.
}
\details{
The `par` argument is a named numeric vector that should have the
following element:
- lambda: the learning rate
- N0: the asymptote
}
\examples{
# write example?
}
