% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaches.R
\name{findStabilizationTrial}
\alias{findStabilizationTrial}
\title{Find index where a timecourse stabilizes using the width of the 95\% CI.}
\usage{
findStabilizationTrial(timecourse, minn = 10, na.rm = TRUE)
}
\arguments{
\item{timecourse}{A vector of values of length>minn representing a timecourse.}

\item{minn}{Minimum number of samples to consider for stabilization (default: 10).}

\item{na.rm}{Logical, whether to remove NA values (default: TRUE).}
}
\value{
Index of the sample where the timecourse stabilizes.
}
\description{
Index where a timecourse stabilizes as determined by the minimum width
of the 95\% confidence interval of the mean from each trial up to the end of the timecourse.
}
\details{
This function implements a method from this paper:
https://doi.org/10.1371/journal.pcbi.1006501
It uses the minimum width of the 95\% confidence intervals of the mean across a decreasing
window. Starting at 1-n, 2-n, 3-n, ..., minn-n, the function calculates the width of
the 95\% confidence interval of the mean for each window. The index of the minimum width
is returned as the point where the timecourse stabilizes.

This works reasonably well (see paper linked above) and is independent of any model
of learning or adaptation.
}
\examples{
# there is example data in this package:
data("tworatedata")

# first we baseline it, and get a median for every trial:
baseline <- function(reachvector,blidx) reachvector - mean(reachvector[blidx], na.rm=TRUE)
tworatedata[,4:ncol(tworatedata)] <- apply(tworatedata[,4:ncol(tworatedata)], 
                                           FUN=baseline, 
                                           MARGIN=c(2), 
                                           blidx=c(17:32))
                                           
# then we only take the rotated phase:
tworatedata <- tworatedata[which(tworatedata$schedule == -30),]                                            

# and plot the first 6 participants as examples:
layout(matrix(c(1,2,3,4,5,6), nrow=3, ncol=2, byrow=TRUE))
for (ppidx in c(4:9)) {
  ppdata <- tworatedata[,ppidx]
  index <- findStabilizationTrial(ppdata, minn=10, na.rm=TRUE)
  plot(ppdata[c(1:index)], type='l', main=colnames(tworatedata)[ppidx], col='red', xlim=c(0,101), ylim=c(-10,40), xlab='trial',ylab='deviation [deg]')
  lines(c(index:length(ppdata)), ppdata[c(index:length(ppdata))], col='blue')
}
}
