% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMacMatlab.R
\name{getMacMatlab}
\alias{getMacMatlab}
\title{Returns the name of the Matlab app of the latest version in the OSX
Applications folder}
\usage{
getMacMatlab()
}
\value{
Latest Matlab application on Mac OSX
}
\description{
Utility function that returns the latest Matlab app (with the highest version
number) in the OSX Applications folder, e.g. "MATLAB_R2014a.app".
}
\author{
Christoph Schmidt <christoph.schmidt@med.uni-jena.de>
}

