% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isvector.R
\name{isvector}
\alias{isvector}
\title{Checks whether input variable is a vector}
\usage{
isvector(obj)
}
\arguments{
\item{obj}{Data whose type is being tested to be a vector (n,1) or (1,n), n>1}
}
\value{
This function returns a boolean indicating whether input obj is a
  vector
}
\description{
An input is considered to be a vector if it has dimensions (n,1) or (1,n),
n>1. Returns TRUE if the input is a vector according to this definition.
Therefore, input that is a one-dimensional 'matrix' in R (is.matrix = TRUE
and is.vector = FALSE) would also be regarded as a vector.
}
\examples{
v <- c(1, 2, 3)
reach:::isvector(v)

reach:::isvector(t(v))

m <- matrix(1:4, 2, 2)
reach:::isvector(m)

v <- matrix(1:4, 4, 1)
reach:::isvector(v)

}
\author{
Christoph Schmidt <christoph.schmidt@med.uni-jena.de>
}

