% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{multiModalFit}
\alias{multiModalFit}
\title{Fit function to optimize an N-modal distributions to given data.}
\usage{
multiModalFit(x, n = 2, points = 9, best = 9)
}
\arguments{
\item{x}{A sequence of numbers to evaluate the probability of.}

\item{n}{The number of normal distributions to consider.}

\item{points}{The number of points to search in each dimension of a search grid.}

\item{best}{Return the parameters for the `best` best fits.}
}
\value{
The best parameters for N-modal distributions for a data set x.
}
\description{
This function is part of a set of functions to fit and
evaluate multi-modal (normal) distribution of data points.
}
\examples{
multiModalFit(x=c(rnorm(50,0,2),rnorm(100,10,4)), n=2, points=7, best=10)
}
