% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{multiModalModel}
\alias{multiModalModel}
\title{Probability of data points given a multi-modal distribution.}
\usage{
multiModalModel(par, x)
}
\arguments{
\item{par}{A named vector with the model parameter (see details).}

\item{x}{A sequence of numbers to evaluate the probability of.}
}
\value{
A vector of probabilities according to the multi modal normal distribution.
}
\description{
This function is part of a set of functions to fit and
evaluate multi modal (normal) distribution of data points.
}
\details{
This is an experimental work-in-progress model that is supposed to fit a
multi-modal distribution. That is: I'm not sure how well it works, and its' a 
little rough around the edges. For one, the output parameters from the optimal 
fit are in a different format from what the other functions expect. Second, the 
output is usually not a probability function that sums to 1. Finally, it might 
not return the actual best fit.

That said, the `par` argument is a data frame (or named list) with columns:
- m: mean
- s: standard deviation
- w: weight
for every normal distribution of the model
}
