% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{odregress}
\alias{odregress}
\title{Orthogonal Distance Regression.}
\usage{
odregress(x, y)
}
\arguments{
\item{x}{A 'predictor' matrix (variables in columns, N observations in rows).}

\item{y}{Column vector (N rows) with 'dependent' variable.}
}
\value{
A list with: `coeff` (coefficients), `yfit` (fitted values), `err` (errors),
`resd` (residuals), `ssq` (sum of squares) and `normal` (normal).
}
\description{
Perform (multiple) orthogonal distance regression.
}
\details{
This function is copied from  the `pracma` package by Hans W. Borchers:
https://CRAN.R-project.org/package=pracma
}
\examples{

}
